requirements.txt
astroid==1.6.5
Django==2.0
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-modeltranslation==0.12.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
Pillow==5.2.0
pylint==1.9.3
pytz==2018.5
six==1.11.0
wrapt==1.10.11
